requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
python-decouple==3.8
gunicorn==23.0.0
drf-spectacular==0.28.0
